airflow-docker/dags/requirements.txt

numpy
pandas
scikit-learn
apache-airflow-providers-apache-kafka